Revoke refresh_token and access_token on logout
When the option revoke_tokens_on_logout is enabled
try to notify the authorization server on logout that previously
obtained refresh and access tokens are no longer needed.

This requires that a revocation endpoint is discoverable.

If the revoke_tokens_on_logout option is enabled but there is
no revocation endpoint supplied, a debug message
will be logged and the logout process continues.
When there is a revocation endpoint supplied but the
authorization server does not respond to a revocation request
with HTTP status code 200, an error message will be logged but
the logout process still continues as if there was no error.