Add randomly-generated state parameter to logout request
This is recommended by ORY Hydra:
https://www.ory.sh/docs/next/hydra/oauth2#logout

Once Hydra has logged out, it redirects to
post_logout_redirect_uri?state={state}. Since the state is randomly
generated, the request cannot be cached by the browser. This means
that post_logout_redirect_uri can be set to a route behind the proxy,
and the user ends up redirected to the login page.