Correct passing of 'symmetric_key' and thus fix HS??? verification
In function openidc_load_and_validate_jwt_id_token(), it was passing the
wrong 'opts.client_secret' instead of 'opts.symmetric_key' as the HMAC
key, and thus caused signature verification failures for HS??? tokens.

Fix the code to pass the correct 'opts.symmetric_key'.

Meanwhile, update the REAME a bit to clarify that the 'symmetric_key'
option is directly passed to HMAC code and thus cannot be encoded.